+++ src/main/resources/db/migration/V16__add_interest_fields_to_loanee_loan_detail.sql
@@ -2,4 +2,4 @@ ALTER TABLE loanee_loan_detail_entity
     ADD COLUMN interest_rate DOUBLE PRECISION DEFAULT 0.0;
 
 ALTER TABLE loanee_loan_detail_entity
-    ADD COLUMN interest_incurred DOUBLE PRECISION DEFAULT 0.0;
+    ADD COLUMN interest_incurred NUMERIC(19, 2) DEFAULT 0.0;
